requirements.txt
Django==4.1.6
python-telegram-bot==13.11
